backend/requirements.txt
snowflake-connector-python[pandas]
fastapi
uvicorn[standard]
python-dotenv
pandas
prophet
scikit-learn
statsmodels
python-multipart
